requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
aiosqlite>=0.20
asyncpg>=0.29
apscheduler>=3.10
httpx>=0.27
pydantic>=2.6
PyYAML>=6.0
python-dotenv>=1.0
google-api-python-client>=2.120
google-auth>=2.28
pytest>=8.0
pytest-asyncio>=0.23
